rename data modules for coffeeType options
