+++ src/components/plan-page/Plan.jsx
@@ -2,7 +2,7 @@ import React, { Suspense, lazy } from 'react';
 import Hero from './hero-section/Hero';
 import PlanSteps from './plan-section/PlanSteps';
 import Options from './options-section/Options';
-import OrderModalContext from './context/OrderModalContext';
+import OrderModalProvider from './context/OrderModalContext';
 import StoreProvider from './context/Context';
 
 const OrderModal = lazy(() => import('./order-modal/OrderModal'));
@@ -11,14 +11,14 @@ const Plan = () => {
   return (
     <>
       <StoreProvider>
-        <OrderModalContext>
+        <OrderModalProvider>
           <Suspense fallback={<div>...</div>}>
             <OrderModal />
           </Suspense>
           <Hero />
           <PlanSteps />
           <Options />
-        </OrderModalContext>
+        </OrderModalProvider>
       </StoreProvider>
     </>
   );
