+++ src/components/plan-page/options-section/_options.module.scss
@@ -263,7 +263,7 @@
 }
 
 .showOption {
-  display: block;
+  // display: block;
 
   @include tabletMd {
     display: grid;
